@@ -1,6 +1,6 @@
 module github.com/apache/skywalking-eyes
 
-go 1.18
+go 1.23
 
 require (
 	github.com/Masterminds/sprig/v3 v3.2.3
